Only moderators are able to create new review comments
We no longer show buttons and forms to users who can't use them.

+++ app/controllers/topics/review_comments_controller.rb
@@ -1,5 +1,6 @@
 class Topics::ReviewCommentsController < ApplicationController
   def create
+    authorize review_comment, :create?
     if review_comment.save
       redirect_to topic_reviews_path(topic), notice: t(".successfully_created")
     else
@@ -8,6 +9,7 @@ class Topics::ReviewCommentsController < ApplicationController
   end
 
   def destroy
+    authorize review_comment, :destroy?
     if review_comment.destroy
       redirect_to topic_reviews_path(topic), notice: t(".successfully_destroyed")
     else
@@ -18,7 +20,7 @@ class Topics::ReviewCommentsController < ApplicationController
   private
 
   def review_comment
-    if params[:id].present?
+    @_review_comment ||= if params[:id].present?
       ReviewComment.find(params[:id])
     else
       ReviewComment.new(body: review_comment_params[:body], author: current_user,
